Refactor: 내부 클래스 RequestLine을 record로 변경
