+++ src/main/java/webserver/request/HttpRequestParser.java
@@ -1,8 +1,8 @@
-package utils;
+package webserver.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.HttpRequest;
+import utils.HttpRequestUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
